# docs/OPENAPI.yaml
...
openapi: 3.1.0
info:
  title: Oblivion API
  description: 'Backend API for controlled data erasure, recovery, residual analysis, assurance and certificate
    verification.


    This document is GENERATED by scripts/gen_openapi.py from the FastAPI application itself. Do not hand-edit it:
    edit the routes and schemas, then regenerate.


    Authentication uses HTTP Bearer with opaque server-side session tokens (Argon2id password hashing, SHA-256 token
    digest stored server-side). These are NOT JWTs. There is no refresh-token endpoint and no cookie based session;
    the bearer token is the only credential transport.'
  version: 0.1.0
paths:
...
  /api/operations:
    post:
      tags:
      - operations
      summary: Create Operation
      description: 'Creates an erasure operation in PENDING_APPROVAL state.

        Does NOT execute destructive deletion upon creation (Approval-gated lifecycle).'
      operationId: create_operation_api_operations_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOperationRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/operations/{operation_id}/approve:
    post:
      tags:
      - operations
      summary: Approve Operation
      description: 'Approves a pending operation subject to Separation of Duties enforcement.

        Requires ADMIN role / operation.approve permission.'
      operationId: approve_operation_api_operations__operation_id__approve_post
      parameters:
      - name: operation_id
        in: path
        required: true
        schema:
          type: string
          title: Operation Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/operations/{operation_id}/execute:
    post:
      tags:
      - operations
      summary: Execute Operation Endpoint
      description: 'Executes an approved operation.

        Requires OPERATOR role / operation.execute permission.

        Enforces that operation was approved prior to execution.'
      operationId: execute_operation_endpoint_api_operations__operation_id__execute_post
      parameters:
      - name: operation_id
        in: path
        required: true
        schema:
          type: string
          title: Operation Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/operations/{operation_id}:
    get:
      tags:
      - operations
      summary: Get Operation
      description: Get operation status and details (requires operation.view permission).
      operationId: get_operation_api_operations__operation_id__get
      parameters:
      - name: operation_id
        in: path
        required: true
        schema:
          type: string
          title: Operation Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/operations/{operation_id}/cancel:
    post:
      tags:
      - operations
      summary: Cancel Operation
      description: Request operation cancellation (requires operation.request permission).
      operationId: cancel_operation_api_operations__operation_id__cancel_post
      parameters:
      - name: operation_id
        in: path
        required: true
        schema:
          type: string
          title: Operation Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/operations/{operation_id}/events:
    get:
      tags:
      - operations
      summary: Get Operation Events
      description: Retrieve audit and lifecycle events for an operation (requires operation.view permission).
      operationId: get_operation_events_api_operations__operation_id__events_get
      parameters:
      - name: operation_id
        in: path
        required: true
        schema:
          type: string
          title: Operation Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OperationEventOut'
                title: Response Get Operation Events Api Operations  Operation Id  Events Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ConfirmationInput:
      properties:
        acknowledged_risk:
          type: boolean
          title: Acknowledged Risk
          description: Must explicitly acknowledge destructive operation risk
      type: object
      required:
      - acknowledged_risk
      title: ConfirmationInput
...
    CreateOperationRequest:
      properties:
        target_id:
          type: string
          minLength: 1
          title: Target Id
        mode:
          type: string
          pattern: ^(COMPLETE_ERASURE|SELECTIVE_PERMANENT|CONTROLLED_RECOVERABLE)$
          title: Mode
        policy_id:
          type: string
          title: Policy Id
          default: ERASURE.LOGICAL.SELECTIVE.V1
          description: 'Must be an allowlisted backend policy. Currently registered: ERASURE.LOGICAL.SELECTIVE.V1,
            ERASURE.LOGICAL.TREE.V1, ERASURE.RECOVERABLE.ENCRYPTED.V1. The backend rejects any other identifier;
            the frontend must not invent policy identifiers.'
        confirmation:
          $ref: '#/components/schemas/ConfirmationInput'
        recovery:
          anyOf:
          - $ref: '#/components/schemas/RecoveryConfig'
          - type: 'null'
      type: object
      required:
      - target_id
      - mode
      - confirmation
      title: CreateOperationRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OperationEventOut:
      properties:
        id:
          type: string
          title: Id
        operation_id:
          type: string
          title: Operation Id
        sequence:
          type: integer
          title: Sequence
        event_type:
          type: string
          title: Event Type
        from_state:
          anyOf:
          - type: string
          - type: 'null'
          title: From State
        to_state:
          anyOf:
          - type: string
          - type: 'null'
          title: To State
        payload:
          anyOf:
          - type: string
          - type: 'null'
          title: Payload
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - operation_id
      - sequence
      - event_type
      - created_at
      title: OperationEventOut
    OperationOut:
      properties:
        id:
          type: string
          title: Id
        target_id:
          type: string
          title: Target Id
        mode:
          type: string
          title: Mode
        state:
          type: string
          title: State
          description: 'Operation lifecycle state. Authoritative values: CREATED, ANALYZING, PENDING_APPROVAL, READY,
            ERASING, VERIFYING, RECOVERY_TEST, RESIDUAL_SCAN, RESIDUAL_ANALYSIS, ASSESSING, CERTIFYING, RECONCILIATION_REQUIRED,
            COMPLETED, PARTIAL, FAILED, INCONCLUSIVE, CANCELLED. Terminal states: COMPLETED, PARTIAL, FAILED, INCONCLUSIVE,
            CANCELLED. Consumers must render PARTIAL and INCONCLUSIVE as distinct outcomes and must never collapse
            them into success.'
        policy_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Policy Id
        progress_percent:
          type: number
          title: Progress Percent
          default: 0.0
        warnings:
          items:
            type: string
          type: array
          title: Warnings
        error:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Error
        actor_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Actor Id
        created_at:
          type: string
          format: date-time
          title: Created At
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
      type: object
      required:
      - id
      - target_id
      - mode
      - state
      - created_at
      title: OperationOut
...
    RecoveryConfig:
      properties:
        retention_seconds:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: Retention Seconds
          default: 86400
      type: object
      title: RecoveryConfig
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: OpaqueSessionToken
      description: 'Opaque session token issued by POST /api/auth/login. Send as ''Authorization: Bearer <token>''.
        The server stores only the SHA-256 digest and revokes it on logout.'
security:
- bearerAuth: []